rabbitmq/consumer_two/requirements.txt
psycopg2-binary
pika==1.3.1